Allow setting OAuth to Read Only
To allow admins to update / migrate the database without affecting
current authorizations, allow putting the extension into a read-only
mode with a config flag.

Bug: 63434

+++ backend/MWOAuthDAO.php
@@ -145,6 +145,9 @@ abstract class MWOAuthDAO implements IDBAccessObject {
 	public function save( DBConnRef $dbw ) {
 		$uniqueId = $this->getIdValue();
 		$idColumn = static::getIdColumn();
+		if ( $dbw->daoReadOnly ) {
+			throw new MWException( get_class( $this ) . ": tried to save while db is read-only" );
+		}
 		if ( $this->daoOrigin === 'db' ) {
 			if ( $this->daoPending ) {
 				wfDebug( get_class( $this ) . ': performing DB update; object changed.' );
@@ -184,6 +187,9 @@ abstract class MWOAuthDAO implements IDBAccessObject {
 	public function delete( DBConnRef $dbw ) {
 		$uniqueId = $this->getIdValue();
 		$idColumn = static::getIdColumn();
+		if ( $dbw->daoReadOnly ) {
+			throw new MWException( get_class( $this ) . ": tried to delete while db is read-only" );
+		}
 		if ( $this->daoOrigin === 'db' ) {
 			$dbw->delete(
 				static::getTable(),
